requirements.txt
Django==2.1.10 
django-pyodbc-azure==2.1.0.0
djangorestframework==3.9.4  
pyodbc==4.0.26 
django-rest-swagger==2.2.0
django-cors-headers==3.0.2
django-filter==2.2.0

